backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*
pydantic-settings==2.7.*
python-dotenv==1.0.*
google-genai>=1.63.0
websockets==14.*
